Corrige intégration de l'édition de compte
