+++ view/sidebar.php
@@ -1,5 +1,5 @@
 <div id="sidebar">
-    <a class="side-button" href="../index.php?target=edition_compte-controller" >Mon compte</a>
+    <a class="side-button" href="../index.php?target=user&page=myinfos" >Mon compte</a>
     <a class="side-button" href="#">Gestion utilisateurs</a>
     <a class="side-button" href="#">Gestion logements</a>
     <a class="side-button" href="#">Tableau de bord</a>
