+++ view/user-compte.php
@@ -1,26 +1,5 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <meta charset="utf-8"/>
-    <title>Prometech</title>
-    <link href="https://use.fontawesome.com/releases/v5.0.8/css/all.css" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
-   
-    <script src="http://code.jquery.com/jquery-3.3.1.min.js"></script>
-    <script src="http://code.jquery.com/ui/1.12.1/jquery-ui.min.js"></script>
-</head>
-<body>
-<?php require_once("header.php"); ?>
-<div id="content-window">
-    <?php require_once("sidebar.php"); ?>
 
-    
-    
- <link rel="stylesheet" type="text/css" href="/public/css/mon_compte.css">
 
- 
-
-	
 <div id="user-compte">
 	
 		<form action="" method="post">
@@ -38,16 +17,9 @@
             <input type="text" name="newprenom" placeholder="<?php echo $userInfos['prenom']?>"   size="25"/>
         </p>
         <br>
-         <!--
-         <p >
-            <label for="adresse">Adresse :</label>
-            <input type="adresse" name="adresse" placeholder="adresse"  size="25"/>
-        </p>
-        <br>
-        -->
         <p >
             <label for="email">Email :</label>
-            <input type="mail" name="newmail"  placeholder="<?php echo $userInfos['email']?>"  size="25"/>
+            <input type="email" name="newmail"  placeholder="<?php echo $userInfos['email']?>"  size="25"/>
         </p>
         <br>
         <p >
@@ -78,19 +50,3 @@
 
 	
 </div>
-    
-    
-    
-</div>
-<!-- Sidebar toggle -->
-<script>
-    $(function () {
-        $("#bars-button ").click(function () {
-            $("#sidebar").toggle(/*'slide', {
-                direction: 'left'
-            }*/);
-        });
-    });
-</script>
-</body>
-</html>
